Script to copy a file
Copies a file and inserts the current date/time partway through the name of the copied file. By copying a file and inserting the current date/time partway through the name of the copied file, a 'backup copy' is made of the original file. This means we can make  amendments to the original file and have a backup copy should we have the need to revert to it. The date and time show when the file was copied and by having a time component, we can make more than one copy per day.