do/until loop and user choice
Test program to store items in the object 'System.Collections.Generic.List'. A do/until loop is then used to limit the user choice from this list of items. So when items are shown in the loop as 0 to 6, for example, the user has to enter an integer value in the range 0 to 6.