Browse (RMAN) log files
Script using the System.Windows.Forms namespace to browse a specific directory structure for (RMAN) log files. The TreeView Class is used to display this hierarchical collection of directories and log files, each represented by a TreeNode. The files could be any log files of course, but the script was written with being able to check RMAN files in mind.