Demonstrates use of a user defined event
Demonstrates a user defined event via the cmdlet 'Register-EngineEvent'. In order to demonstrate using this event, a 'foreach' loop is executed and when the loop variable reaches a certain figure, (4 in this example), the event is fired by the 'New-Event' cmdlet. On firing the event, a user written scriptblock is executed in order to show that something has happened.