Code improvements
Made significant changes to the scriptblock which looks for a file when I don't where it is located. Errors created during the running of this scriptblock are written to variable $Error and by looking at this variable, we can see what errors have occurred whilst the scriptblock has been running. Usually, the error encountered is 'UnauthorizedAccessException'.