List of Camborne Trevithick Day's
Camborne Trevithick Day, always held on the last Saturday of April, celebrates the engineering and mining history of the Camborne area, and is dedicated to Richard Trevithick, pioneer of high pressure steam power and inventor of Road and Railway
Locomotives.

This program lists Camborne Trevithick Day for the current year plus the next four (ie, five years in total). Depending on when you run this program, Camborne Trevithick Day for the current year may well have passed.