Code improvements
Demonstrate use of background job event handling
Event handling. Demonstration file showing how to register for an event when submitting a background job. When the event is raised, an action block ScriptBlock informs us the job has completed. This saves us having to check upon the status of the job every few minutes.