Init version
Demonstration file created by Jeffery Hicks, (SAPIEN Technologies) to display a graphical directory tree using PowerShell and Windows forms. Folders with 0 files and 0 subfolders will be displayed in magenta. This does not count hidden files or folders.
The program makes use of the .NET classes 'System.Windows.Forms.TreeView' and 'System.Windows.Forms.TreeNode' and is the Java equivalent of the Java class 'javax.swing.JTree'. This is a control that displays a hierarchical collection of labeled items, each represented by a TreeNode. ie, in a tree fashion.