Demonstration program of a TreeView and TreeNode
Init version. Demonstration program showing the use of a TreeView and TreeNode. The program also uses a FlowLayoutPanel panel to hold some buttons which react to the 'FlowLayoutPanel' event by the use of a script block. This program is the PowerShell version of SimpleTree.java which uses javax.swing.JTree. Admittedly, the programs doesn't really do anything. It's more of a demonstration program showing how to do something.