Refactored the code
Changed the program parameter name to "Path" as can be found in a lot of PowerShell cmdlets. In addition, did some type checking of the variable type passed in. When the type checking is complete, the variable will be of type 'System.IO.FileInfo'. Updated to reflect the use of Java SDK version 21.